fixed the usleep and time calculation now it uses mili instead of micro seconds and now i free all the allocated memory

+++ codexion.h
@@ -56,6 +56,6 @@ void				compile(t_coder *coder);
 void				debug(t_coder *coder);
 void				refactor(t_coder *coder);
 void				*run_coder(void *arg);
-long long			get_time(long long start);
+long				get_time(long start);
 
 #endif
